# docs/openapi.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "VulnProof application API",
    "version": "0.1.0",
    "description": "Generated from implemented request schemas and checked against registered routes. Arc Testnet only for live funds. Fixture verification uses FIXTURE_ONLY and TRUSTED_SERVICE. Private report downloads and internal service routes are separate services. Response projections without a dedicated DTO schema remain open objects; this contract does not promise undocumented fields."
  },
  "servers": [
    {
      "url": "/",
      "description": "The configured API origin"
    }
  ],
  "paths": {
...
    "/api/v1/organizations/{id}": {
      "get": {
        "operationId": "organization",
        "summary": "Read organization settings and members",
        "description": "Current organization member. All response metadata uses Cache-Control: no-store. Custom schema refinements and current role, wallet, source, and chain checks remain server requirements.",
        "security": [
          {
            "PrivyBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request completed. Financial state is final only where explicitly recorded as final.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "description": "Authorized resource or operation metadata. Additional fields depend on the resource and its saved state. See the route implementation for the full current projection."
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "410": {
            "$ref": "#/components/responses/ApiError"
          },
          "413": {
            "$ref": "#/components/responses/ApiError"
          },
          "415": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
...
    "/api/v1/organizations/{id}/wallets": {
      "get": {
        "operationId": "treasuryWallets",
        "summary": "Read organization wallet setup records",
        "description": "Current organization OWNER or TREASURY. All response metadata uses Cache-Control: no-store. Custom schema refinements and current role, wallet, source, and chain checks remain server requirements.",
        "security": [
          {
            "PrivyBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request completed. Financial state is final only where explicitly recorded as final.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "description": "Authorized resource or operation metadata. Additional fields depend on the resource and its saved state. See the route implementation for the full current projection."
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "410": {
            "$ref": "#/components/responses/ApiError"
          },
          "413": {
            "$ref": "#/components/responses/ApiError"
          },
          "415": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      },
      "post": {
        "operationId": "createTreasuryWallet",
        "summary": "Queue organization wallet setup with an amount cap",
        "description": "Current organization OWNER. All response metadata uses Cache-Control: no-store. Custom schema refinements and current role, wallet, source, and chain checks remain server requirements.",
        "security": [
          {
            "PrivyBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Reuse this key only for identical input. The server keeps the request hash for at least 24 hours. A changed request returns 409.",
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "minLength": 16,
              "maxLength": 128,
              "pattern": "^[a-zA-Z0-9:_-]+$"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "The operation is queued. This is not proof of final chain settlement.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "description": "Authorized resource or operation metadata. Additional fields depend on the resource and its saved state. See the route implementation for the full current projection."
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "410": {
            "$ref": "#/components/responses/ApiError"
          },
          "413": {
            "$ref": "#/components/responses/ApiError"
          },
          "415": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Request_treasurySetup"
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{id}/wallets/{walletId}/balance": {
      "get": {
        "operationId": "treasuryBalance",
        "summary": "Read the organization wallet test USDC balance",
        "description": "Current organization OWNER or TREASURY. All response metadata uses Cache-Control: no-store. Custom schema refinements and current role, wallet, source, and chain checks remain server requirements.",
        "security": [
          {
            "PrivyBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          {
            "name": "walletId",
            "in": "path",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request completed. Financial state is final only where explicitly recorded as final.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "description": "Authorized resource or operation metadata. Additional fields depend on the resource and its saved state. See the route implementation for the full current projection."
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "410": {
            "$ref": "#/components/responses/ApiError"
          },
          "413": {
            "$ref": "#/components/responses/ApiError"
          },
          "415": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
...
    "/api/v1/rpc/arc": {
      "post": {
        "operationId": "arcRpc",
        "summary": "Proxy bounded Arc wallet RPC calls",
        "description": "Public. Reads are allowlisted. Signed sends require an exact saved Privy transfer intent. All response metadata uses Cache-Control: no-store. Custom schema refinements and current role, wallet, source, and chain checks remain server requirements.",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request completed. Financial state is final only where explicitly recorded as final.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/RpcResponse"
                    },
                    {
                      "type": "array",
                      "minItems": 1,
                      "maxItems": 10,
                      "items": {
                        "$ref": "#/components/schemas/RpcResponse"
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiError"
          },
          "401": {
            "$ref": "#/components/responses/ApiError"
          },
          "403": {
            "$ref": "#/components/responses/ApiError"
          },
          "404": {
            "$ref": "#/components/responses/ApiError"
          },
          "409": {
            "$ref": "#/components/responses/ApiError"
          },
          "410": {
            "$ref": "#/components/responses/ApiError"
          },
          "413": {
            "$ref": "#/components/responses/ApiError"
          },
          "415": {
            "$ref": "#/components/responses/ApiError"
          },
          "429": {
            "$ref": "#/components/responses/ApiError"
          },
          "503": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Request_arcRpc"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "ApiError": {
        "description": "The request failed. The error code is stable. The message contains no private provider details.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "PrivyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Privy access token",
        "description": "The API verifies the token and rechecks current database membership. Local test identities work only with the explicit local adapter."
      }
    },
    "schemas": {
      "ApiError": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              }
            },
            "required": [
              "code",
              "message",
              "requestId",
              "retryable"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "RpcResponse": {
        "type": "object",
        "required": [
          "jsonrpc",
          "id"
        ],
        "properties": {
          "jsonrpc": {
            "const": "2.0"
          },
          "id": {
            "type": [
              "string",
              "integer",
              "null"
            ]
          },
          "result": {},
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              }
            }
          }
        },
        "oneOf": [
          {
            "required": [
              "result"
            ]
          },
          {
            "required": [
              "error"
            ]
          }
        ]
      },
...
      "Request_treasurySetup": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "maxPerAction": {
            "type": "string",
            "maxLength": 78,
            "pattern": "^(0|[1-9][0-9]*)$"
          }
        },
        "required": [
          "maxPerAction"
        ],
        "additionalProperties": false,
        "description": "maxPerAction is from 1 to 100000000 base units, inclusive. This is at most 100 test USDC."
      },
...
      "Request_arcRpc": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "anyOf": [
          {
            "type": "object",
            "properties": {
              "jsonrpc": {
                "type": "string",
                "const": "2.0"
              },
              "id": {
                "anyOf": [
                  {
                    "type": "string",
                    "maxLength": 100
                  },
                  {
                    "type": "integer",
                    "minimum": -9007199254740991,
                    "maximum": 9007199254740991
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "method": {
                "type": "string"
              },
              "params": {
                "default": [],
                "maxItems": 5,
                "type": "array",
                "items": {}
              }
            },
            "required": [
              "jsonrpc",
              "id",
              "method"
            ],
            "additionalProperties": false
          },
          {
            "minItems": 1,
            "maxItems": 10,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "jsonrpc": {
                  "type": "string",
                  "const": "2.0"
                },
                "id": {
                  "anyOf": [
                    {
                      "type": "string",
                      "maxLength": 100
                    },
                    {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "method": {
                  "type": "string"
                },
                "params": {
                  "default": [],
                  "maxItems": 5,
                  "type": "array",
                  "items": {}
                }
              },
              "required": [
                "jsonrpc",
                "id",
                "method"
              ],
              "additionalProperties": false
            }
          }
        ],
        "description": "One JSON-RPC request or a batch of at most ten requests. The route limits the whole body to 50000 bytes. Read methods are allowlisted. Contract calls use the fixed Arc USDC address. Signed sends and gas estimates must match saved user transfer intents. Rejected methods return a JSON-RPC error inside HTTP 200."
      }
    }
  }
}